EditDialog: stop floating actionButtons in mobile
