+++ src/components/FeaturePanel/EditDialog/EditDialog.tsx
@@ -1,18 +1,8 @@
 import React, { useState } from 'react';
-import Dialog from '@material-ui/core/Dialog';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import DialogActions from '@material-ui/core/DialogActions';
-import Button from '@material-ui/core/Button';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
-import useTheme from '@material-ui/core/styles/useTheme';
-import DialogContent from '@material-ui/core/DialogContent';
-import { CircularProgress } from '@material-ui/core';
-import styled from 'styled-components';
 import { useRouter } from 'next/router';
 import { useToggleState } from '../../helpers';
-import { Feature, FeatureTags } from '../../../services/types';
-import { createNoteText } from './createNoteText';
-import { insertOsmNote } from '../../../services/osmApi';
+import { Feature } from '../../../services/types';
 import { MajorKeysEditor } from './MajorKeysEditor';
 import {
   ChangeLocationEditor,
@@ -24,40 +14,19 @@ import {
 } from './components';
 import { OtherTagsEditor } from './OtherTagsEditor';
 import { SuccessContent } from './SuccessContent';
-import { addOsmFeature, editOsmFeature } from '../../../services/osmApiAuth';
 import { useOsmAuthContext } from '../../utils/OsmAuthContext';
 import { t } from '../../../services/intl';
 import Maki from '../../utils/Maki';
 import { FeatureTypeSelect } from './FeatureTypeSelect';
-import { getLabel } from '../../../helpers/featureLabel';
-
-const useIsFullScreen = () => {
-  const theme = useTheme();
-  return useMediaQuery(theme.breakpoints.down('sm'));
-};
-
-const useTagsState = (
-  initialTags: FeatureTags,
-): [FeatureTags, (k: string, v: string) => void] => {
-  const [tags, setTags] = useState(initialTags);
-  const setTag = (k, v) => setTags((state) => ({ ...state, [k]: v }));
-  return [tags, setTag];
-};
-
-const useGetDialogTitle = (isAddPlace, isUndelete, feature) => {
-  const { loggedIn } = useOsmAuthContext();
-  if (isAddPlace) return t('editdialog.add_heading');
-  if (isUndelete) return t('editdialog.undelete_heading');
-  if (!loggedIn)
-    return `${t('editdialog.suggest_heading')} ${getLabel(feature)}`;
-  return `${t('editdialog.edit_heading')} ${getLabel(feature)}`;
-};
-
-const StyledDialog = styled(Dialog)`
-  .MuiDialog-container.MuiDialog-scrollPaper {
-    align-items: start;
-  }
-`;
+import { ActionButtons } from './ActionButtons';
+import {
+  ResponsiveDialogContent,
+  StyledDialog,
+  useGetDialogTitle,
+  useIsFullScreen,
+  useTagsState,
+} from './utils';
+import { saveDialog } from './saveDialog';
 
 interface Props {
   feature: Feature;
@@ -67,50 +36,6 @@ interface Props {
   isAddPlace: boolean;
   isUndelete: boolean;
 }
-
-const saveDialog = ({
-  feature,
-  typeTag,
-  tags,
-  tmpNewTag,
-  cancelled,
-  location,
-  comment,
-  loggedIn,
-  setLoading,
-  setSuccessInfo,
-  isUndelete,
-}) => {
-  const tagsWithType = typeTag
-    ? { [typeTag.key]: typeTag.value, ...tags }
-    : tags;
-  const allTags = { ...tagsWithType, ...tmpNewTag }; // we need to send also unsubmitted new tag
-  const noteText = createNoteText(
-    feature,
-    allTags,
-    cancelled,
-    location,
-    comment,
-    isUndelete,
-  );
-  if (noteText == null) {
-    // TODO we need better check that this ... formik?
-    alert(t('editdialog.changes_needed')); // eslint-disable-line no-alert
-    return;
-  }
-
-  setLoading(true);
-  const promise = loggedIn
-    ? feature.point
-      ? addOsmFeature(feature, comment, allTags)
-      : editOsmFeature(feature, comment, allTags, cancelled)
-    : insertOsmNote(feature.center, noteText);
-
-  promise.then(setSuccessInfo, (err) => {
-    console.error(err); // eslint-disable-line no-console
-    setTimeout(() => setLoading(false), 1000);
-  });
-};
 
 export const EditDialog = ({
   feature,
@@ -156,6 +81,15 @@ export const EditDialog = ({
 
   const dialogTitle = useGetDialogTitle(isAddPlace, isUndelete, feature);
 
+  const actionButtons = (
+    <ActionButtons
+      loading={loading}
+      onClose={onClose}
+      handleSave={handleSave}
+      cancelled={cancelled}
+    />
+  );
+
   return (
     <StyledDialog
       fullScreen={fullScreen}
@@ -169,60 +103,37 @@ export const EditDialog = ({
       {successInfo ? (
         <SuccessContent successInfo={successInfo} handleClose={onClose} />
       ) : (
-        <>
-          <DialogContent dividers>
-            <form autoComplete="off" onSubmit={(e) => e.preventDefault()}>
-              {false && (
-                <FeatureTypeSelect type={typeTag} setType={setTypeTag} />
-              )}
-              <MajorKeysEditor
-                tags={tags}
-                setTag={setTag}
-                focusTag={focusTag}
-              />
+        <ResponsiveDialogContent actionButtons={actionButtons}>
+          <form autoComplete="off" onSubmit={(e) => e.preventDefault()}>
+            {false && <FeatureTypeSelect type={typeTag} setType={setTypeTag} />}
+            <MajorKeysEditor tags={tags} setTag={setTag} focusTag={focusTag} />
 
-              {!isAddPlace && !isUndelete && (
-                <>
-                  <DialogHeading>
-                    {t('editdialog.options_heading')}
-                  </DialogHeading>
-                  <PlaceCancelledToggle
-                    cancelled={cancelled}
-                    toggle={toggleCancelled}
-                  />
-                  <ChangeLocationEditor
-                    location={location}
-                    setLocation={setLocation}
-                    feature={feature}
-                  />
-                </>
-              )}
-              <ContributionInfoBox />
-              <CommentField comment={comment} setComment={setComment} />
-              <OtherTagsEditor
-                tags={tags}
-                setTag={setTag}
-                focusTag={focusTag}
-                setTmpNewTag={setTmpNewTag}
-              />
+            {!isAddPlace && !isUndelete && (
+              <>
+                <DialogHeading>{t('editdialog.options_heading')}</DialogHeading>
+                <PlaceCancelledToggle
+                  cancelled={cancelled}
+                  toggle={toggleCancelled}
+                />
+                <ChangeLocationEditor
+                  location={location}
+                  setLocation={setLocation}
+                  feature={feature}
+                />
+              </>
+            )}
+            <ContributionInfoBox />
+            <CommentField comment={comment} setComment={setComment} />
+            <OtherTagsEditor
+              tags={tags}
+              setTag={setTag}
+              focusTag={focusTag}
+              setTmpNewTag={setTmpNewTag}
+            />
 
-              <OsmLogin />
-            </form>
-          </DialogContent>
-          <DialogActions>
-            {loading && <CircularProgress size={20} />}
-            <Button onClick={onClose} color="primary">
-              {t('editdialog.cancel_button')}
-            </Button>
-            <Button onClick={handleSave} color="primary" variant="contained">
-              {loggedIn
-                ? cancelled
-                  ? t('editdialog.save_button_delete')
-                  : t('editdialog.save_button_edit')
-                : t('editdialog.save_button_note')}
-            </Button>
-          </DialogActions>
-        </>
+            <OsmLogin />
+          </form>
+        </ResponsiveDialogContent>
       )}
     </StyledDialog>
   );
